targets/litex_sim: get the conf_soc parameters right
BaseSoC() is eating-up the "serial" interface from the
list of resources available of the platform. Which
means a Platform() cannot be it cannot be passed twice
to a BaseSoC().

This commit uses a new Platform() infocation to work
around this for the conf_soc used as a preflight to
get the memory parameters, before the real BaseSoC()
is invoked.